some how linq select account works

+++ ClearingFramework/dbBind/pageDatabase/Account.cs
@@ -16,7 +16,7 @@ namespace ClearingFramework.dbBind.pageDatabase
     {
         public long id { get; set; }
         public long memberid { get; set; }
-        public Nullable<long> accNum { get; set; }
+        public Nullable<int> accNum { get; set; }
         public string accType { get; set; }
         public string LinkAcc { get; set; }
         public Nullable<decimal> collateral { get; set; }
